# backend/api-schemas/api-swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/v1/auth/signup":{"post":{"operationId":"AuthController_signUp","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/signin":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninAuthDTO"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/v1/users/create":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"]}},"/api/v1/users/all":{"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/v1/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]},"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/v1/workspaces/create":{"post":{"operationId":"WorkspacesController_create","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"logo":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":""}},"tags":["Workspaces"]}},"/api/v1/workspaces/all":{"get":{"operationId":"WorkspacesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Workspaces"]}},"/api/v1/workspaces/{id}":{"get":{"operationId":"WorkspacesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Workspaces"]},"delete":{"operationId":"WorkspacesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Workspaces"]}},"/api/v1/workspaces/update/{id}":{"patch":{"operationId":"WorkspacesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateWorkspaceDto"}}}},"responses":{"200":{"description":""}},"tags":["Workspaces"]}},"/api/tasks/create":{"post":{"operationId":"TasksController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskDto"}}}},"responses":{"201":{"description":""}}}},"/api/tasks":{"get":{"operationId":"TasksController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/api/tasks/{id}":{"get":{"operationId":"TasksController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"TasksController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTaskDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"TasksController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/v1/projects/create":{"post":{"operationId":"ProjectsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectDto"}}}},"responses":{"201":{"description":""}},"tags":["projects"]}},"/api/v1/projects/all/{workspaceId}":{"get":{"operationId":"ProjectsController_findAll","parameters":[{"name":"workspaceId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["projects"]}},"/api/v1/projects/{id}":{"get":{"operationId":"ProjectsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["projects"]},"patch":{"operationId":"ProjectsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectDto"}}}},"responses":{"200":{"description":""}},"tags":["projects"]},"delete":{"operationId":"ProjectsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["projects"]}},"/api/v1/members/invite":{"post":{"operationId":"MembersController_invite","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InviteMemberDto"}}}},"responses":{"201":{"description":""}},"tags":["Members"]}},"/api/v1/members/respond-invitation":{"post":{"operationId":"MembersController_respondToInvitation","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponsdInvitationDto"}}}},"responses":{"201":{"description":""}},"tags":["Members"]}},"/api/v1/members/all/{workspaceId}":{"get":{"operationId":"MembersController_findAllWorkspaceMembers","summary":"Get all members of a workspace","parameters":[{"name":"workspaceId","required":true,"in":"path","description":"The unique ID of the workspace","schema":{"example":"6746be277b3137ce276164ad","type":"string"}}],"responses":{"200":{"description":""}},"tags":["Members"]}},"/api/v1/members/{userId}":{"get":{"operationId":"MembersController_findallMemberships","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Members"]}},"/api/v1/members/update/{memberId}":{"patch":{"operationId":"MembersController_updateMemerRole","parameters":[{"name":"memberId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMemberDto"}}}},"responses":{"200":{"description":""}},"tags":["Members"]}},"/api/v1/members/{id}":{"delete":{"operationId":"MembersController_removeMemer","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Members"]}}},"info":{"title":"Jirox API","description":"Jira Clone API","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"SigninAuthDTO":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","description":"User Role","default":"USER"}},"required":["email","password","role"]},"UpdateUserDto":{"type":"object","properties":{"name":{"type":"string"},"image":{"type":"string"}},"required":["name","image"]},"UpdateWorkspaceDto":{"type":"object","properties":{}},"CreateTaskDto":{"type":"object","properties":{}},"UpdateTaskDto":{"type":"object","properties":{}},"CreateProjectDto":{"type":"object","properties":{}},"UpdateProjectDto":{"type":"object","properties":{}},"InviteMemberDto":{"type":"object","properties":{}},"ResponsdInvitationDto":{"type":"object","properties":{}},"UpdateMemberDto":{"type":"object","properties":{}}}}}